link to printf function in codegen
this is superrrr primitive and needs refactoring. I added the ability to call printf, but its rather hacky. We need a list of functions to be added to the global environment when they are called(printf is one of them). We also need support for strings in the AST and codegen. I think things get a little tricky with strings... Not sure though, but either way the hack to introduce printf needs abstraction