progress towards a core IR
This is almost done. Up to this point, we have a core IR,
formal transformations from AST->Core IR, core IR data type,
a 3/4 implemented transformation, a 3/4 implemented codegen,
and an implemented driver.

The hard part left is to finish codegen. Apparently, LLVM uses
SSA. This means mutation is not allowed, but LLVM has ways
to address this.