fix bytearr and implement array access
This is a major step towards getting this complete.
We have a byte array which we can assign an initial value
and access elements. I do think the byte array AST and core IR
must be looked at again. Im not sure if a string is the best
way to initialize it.