implement for loops
This gets for loops working. Some notes about them --
The step value is always set to 1. We may want to change
this in the future. Also, the AST was changed so that the
high and low are expressions. This needs to be tested
more so that the expression is a primitive number, not
that it evaluates to a number.