fix function calls within constantc
constantc function calls were messed up because right now
everything is pass by reference. The code was passing
the pointer to the value. In the case of byte arrays,
it was passing a pointer to the pointer causing type
errors. This was fixed by keeping track of function
arg types and choosing when to use pointers
appropriately.