implement most of the syntax
This covers most of the syntax, but it is not robust.
It needs to be heavily tested, comments need to be
implemented in the lexer, error message positions need
to be passed to the AST correctly, and bytearr declarations
need to be implemented. Im sure I am missing a few things,
but these are the big ones.