implement the label inference algorithm
This commit implements label inference. It does this by
introducing labeled types which replace the previous type.
Labeled types are simply types with an optional Public/Private
label. The algorithm follows a few rules for inference. First,
literals are always public. Second, function args are never
inferred. They are private by default. Third, at the end of the
typecheck, if a label is ambiguous, it is set to private.
This has no effect as of right now, but it will when we
use labels in the core language for optimizations. Finally,
public values can flow to private values, but private
values cannot flow to public values.