initial stab at ref/out params
This allows any argument to be a ref. A ref is a writable
box for a variable. The interface with C is that a pointer
is passed. This is useful because it assumes the value
is set and the caller has some need for an updated value.

This allows starts the code for out params. An out param is
slightly different than a ref in that it is not assumed to be
set? We need to figure this out exactly, but right now it
behaves the same as a ref