add more int types
This is rather hacky and not yet completely flushed out.
The idea was to add multiple types of ints(32 bits, 16, bits,
and 8 bits). Easy. The complication occurs with operator
overloading and implicit casting. Things get funky in codegen.

In codegen:
The LHS(declarations) passes a type context to expression codegen.
This is used to tell the expression what is expected on the LHS.
This is really only important when the type is inferred through
a literal. Since the literal knows what it will be cast to, it
just does it immediately, preventing any issues later on
(casting a 32 bit int to 16 bits, etc).

During binary operations, the LHS and RHS are casted together.
Simple enough.

In the variable declaration, the RHS value is computed, then
cast to match the LHS.

This is not really tested with arrays. Everything is assumed to
be 32 bits. I semi-intentionally ignored them since they will
be recieving a lot of attention coming up and most of that code
will likely change.