add BoolMask type
This type is required for codegen. It essentially allows us
to combine boolean values with anything by extending
or truncating the boolean. This lets us use the special,
injected variables with both signed and unsigned numbers.
Prior to this, these injected variables had to be either
signed or unsigned. This would cause issues with codegen
because it had to work with both types, thus would fail.
Now we just unify it to whatever it needs to be and this
always works because it is either TRUE or FALSE.

Future work: BoolMasks are represented as both i8's and
i32's. They are i8's when used inside a function. They are
i32's when they are returned from a function. The reason for
this is backwards compatibility. The i8 was a mini optimization
(whether necessary is another discussion). I believe that
they should be updated to be i1's in the future. This would
mean they take up less space and will be consistent across the
board. I am unsure if the injected variables will work as
expected if they are i1's though so more thought needs to
occur here.