Progagate positions to the core IR
We need to know positions all the way up to where we generate
LLVM. This way, we can handle any exceptions with accuracy.
There are some problems with our positions though. Mainly that
we inject a lot of code. We lose accuracy for injected code.
Meaning, we give a position, but it may point to valid code.
To solve this, we should have two types of positions: Real
and Injected.