Use errors from err.ml
We want to aggregate all of the errors in a single place.
This way we can catch each error individually at the end.
It will also make it easy to assign each error an error code.
This will make it possible to test invalid constanc programs.
Furthermore, this makes it really nice to throw errors. You
simply choose an error, pass along the necessary information
and the error module handles everything else(message generation,
reporting, etc.)