Improve error messages
Previously, the error messages had only displayed the error code.
The developer would then have to look up the error code to see
what was happening. Now, we print the a real message designed
to assist in giving clues to the developer as to what went wrong.
This is not perfect, but it will change over time. The important
thing is that errors are centralized, simple, and easy to update.