Update syntax
This is the updated syntax for SOSP. This is part of a
bigger WIP. The changes here show how some parts of the
language were changed/refined and how new parts were added.
This only updates the top level AST with the parser and lexer.
Changes to come include propagating the new AST all the way
through the pipeline, but each major step will have its own
PR.