WIP: REPL exprs
We can now run a REPL over expressions. This should
be expanded to add function declarations in the future
so people can write a function and call it. Also, we
should be able to load a file and call those functions.

To use this, build, then run ./facti.byte. Expressions
are delimited by ;;