implement use of arrays from decs and passed as args
So before, the GEP instruction was assuming the type of
the value was an array. Unfortunately, arrays are passed
as a pointer to type. This causes the GEP instruction to
segfault due to the type differences.

This change keeps track of the arrays that are pointers
(passed as an argument) or declared as an array (in the
function body) in order to generate the proper GEP
instruction.