implement early return inside if-statements
Previously, the if-statement codegen would always merge
the then and else basic blocks. This causes an LLVM
termination error when one of the blocks has an early
return. This change simply checks whether a basic block
has returned. If it has, then it does not branch to the
merge block.