pass left hand type to arrgen from arrdec
While generating arrays from the array declaration, we were
passing the array expression type. The actual type of this
cannot always be derived from the expression (e.g., in the
case of arrviews). Thus, we need to pass the type from the
declaration -- which is the type the array needs to fit
regardless. By passing the expression type, it guesses to
fit a 32 bit width.

Fixes #45