fix arrview pass bug
Arrviews are unique on the codegen side because it is just a pointer.
The old code was treating it as an llvm array so the pointer was
being casted improperly causing it to point to the wrong place in
memory. This change checks that it only casts array types and loads
pointer types when passing an array via a function call.