WIP: add optimization checker
When we run optimizations we need to be sure a timing channel isnt
introduced into the IR. This commit makes progress towards this by
first adding prefixes to temp names. This allows us to analyze the IR
to determine if secret data can be leaked. Second, it implements a
label inferencer. This makes it possible to determine labels of temps
generated by an optimization so that the checkers can analyze the
optimized code. Third, it creates checkers for branching, reads,
and writes. These checkers respectively ensure that branching is
only done on public values, public values dont read secret values,
and secret values don't write to public values.