start fact integration with the liquid haskell site
The liquid haskell site has specific requirements for
how to communicate with the compiler. This commit adds
a -json flag that will communicate the information in a
way that the site can understand.

Right now, this will send an error message every time we
try to compile. Next steps are to actually compile the
passed file.